# core/vm: use uint256 in EVM implementation