--- a/core/vm/eips.go
+++ b/core/vm/eips.go
@@ -18,6 +18,7 @@ package vm
 
 import (
 	"fmt"
+	"github.com/holiman/uint256"
 
 	"github.com/ethereum/go-ethereum/params"
 )
@@ -63,8 +64,7 @@ func enable1884(jt *JumpTable) {
 }
 
 func opSelfBalance(pc *uint64, interpreter *EVMInterpreter, callContext *callCtx) ([]byte, error) {
-	balance := interpreter.intPool.get()
-	balance.SetFromBig(interpreter.evm.StateDB.GetBalance(callContext.contract.Address()))
+	balance, _ := uint256.NewFromBig(interpreter.evm.StateDB.GetBalance(callContext.contract.Address()))
 	callContext.stack.push(balance)
 	return nil, nil
 }
@@ -84,8 +84,7 @@ func enable1344(jt *JumpTable) {
 
 // opChainID implements CHAINID opcode
 func opChainID(pc *uint64, interpreter *EVMInterpreter, callContext *callCtx) ([]byte, error) {
-	chainId := interpreter.intPool.get()
-	chainId.SetFromBig(interpreter.evm.chainConfig.ChainID)
+	chainId, _ := uint256.NewFromBig(interpreter.evm.chainConfig.ChainID)
 	callContext.stack.push(chainId)
 	return nil, nil
 }