… core/vm/instructions_test.go
… core/vm/int_pool_verifier.go
… core/vm/int_pool_verifier_empty.go
… core/vm/interpreter.go
… core/vm/intpool.go
… core/vm/intpool_test.go
… core/vm/logger.go
… core/vm/logger_json.go
… core/vm/logger_test.go
… core/vm/memory.go
… core/vm/stack.go
… eth/tracers/tracer.go
… go.mod
… go.sum