Continuation of #20601 by @holiman.

> This PR is another experiment at integrating the fixed-256 bit math library, instead of using the native go big.Int arbitrary precision big number library.
>
> Theoretically, it could make a difference, but historically we haven't seen any big differences. We have some more advanced metrics now, and both @gballet and @chfast has been curious about giving it another go, so let's do a benchmark when we have a couple of machines available.
>
> See #18191 for earlier numbers.

This one uses `uint256@v1.0.0` and removes intpool.

The git history is not perfect, let me know if I should work more on it.