@holiman 2020-03-20
I pushed a loop benchmark as the first commit. 
```
benchstat loop_v0.txt loop_v4.txt 
name          old time/op    new time/op    delta
SimpleLoop-6     198ms ± 5%     145ms ± 1%  -26.52%  (p=0.008 n=5+5)

name          old alloc/op   new alloc/op   delta
SimpleLoop-6    51.6kB ± 0%    76.1kB ± 0%  +47.35%  (p=0.008 n=5+5)

name          old allocs/op  new allocs/op  delta
SimpleLoop-6       110 ± 0%       109 ± 0%     ~     (p=0.079 n=4+5)
[user@work vm]$ 
```