@holiman 2020-03-20
This is a substantial change, on a loop which does not access state, but performs some rather common ops, like jump, jumpdest, sub, swap, dup, it shaves of 25%.