@holiman 2020-03-20
I pushed another commit, which improves the lookup of a jumpdest analysis and heavily improves the performance of `JUMP`/`JUMPI`. We might want to cherry-pick that to a separate pr. Here's the bench diff when applying the `core/vm: improve jumpdest lookup` commit

```
loop_v4.txt loop_v5.txt 
name          old time/op    new time/op    delta
SimpleLoop-6     145ms ± 1%     115ms ± 1%  -20.86%  (p=0.008 n=5+5)

name          old alloc/op   new alloc/op   delta
SimpleLoop-6    76.1kB ± 0%    76.1kB ± 0%     ~     (p=0.635 n=5+4)

name          old allocs/op  new allocs/op  delta
SimpleLoop-6       109 ± 0%       109 ± 0%     ~     (all equal)
```