@holiman 2020-03-24
Some charts. This PR is ahead by about 45 minutes, after a couple of days. 

![Screenshot_2020-03-24 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/77411410-ace21b80-6dbc-11ea-991d-3263cf240ff8.png)