@holiman 2020-03-24
This PR has consistently lower allocation count 
![Screenshot_2020-03-24 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/77411494-d1d68e80-6dbc-11ea-99b6-574728235e96.png)