@holiman 2020-05-21
Funky that travis got this on one of the builds:
```
--- FAIL: TestSingleSideRangeProof (0.05s)

    proof_test.go:426: Expected no error, got the fork point must be a fullnode

{op: 6, key: common.Hex2Bytes(""), value: common.Hex2Bytes("")}, // step 0

{op: 6, key: common.Hex2Bytes(""), value: common.Hex2Bytes("")}, // step 1

{op: 0, key: common.Hex2Bytes("d51b182b95d677e5f1c82508c0228de96b73092d78ce78b2230cd948674f66fd1483bd"), value: common.Hex2Bytes("0000000000000002")}, // step 2
```

..?