@chfast 2020-05-25
> Might be a good idea to also pull in this one: [ledgerwatch/turbo-geth@7f9557f](https://github.com/ledgerwatch/turbo-geth/commit/7f9557f35db532cab63054c33cadd1ccf920c69e)

I was considering doing the same change, but there is probably a way to eliminate the need for some parts of the `Safe` interface.