@chfast 2020-05-25
I can fix `uint256` build for Go 1.12 if that's desired.