@holiman 2020-05-25
> I can fix uint256 build for Go 1.12 if that's desired.

It's definitely not needed by geth -- due to the new error handling introduced in 1.13, we're going to deprecate 1.12 anyway. Both the new snapshot sync implementation and https://github.com/ethereum/go-ethereum/pull/21067 deprecates 1.13