@holiman 2020-05-25
Ok ,benchmark is started. https://geth-bench.ethdevops.io/d/Jpk-Be5Wk/dual-geth?orgId=1&var-exp=bench01&var-master=bench02&var-percentile=50&from=1590434109271&to=now . 

```
May 25 21:14:33 bench01.ethdevops.io geth INFO [05-25|19:14:33.545] Starting peer-to-peer node instance=Geth/v1.9.15-unstable-30b26f83-20200525/linux-amd64/go1.14.3
May 25 21:14:40 bench02.ethdevops.io geth INFO [05-25|19:14:40.123] Starting peer-to-peer node instance=Geth/v1.9.15-unstable-dbf6b8a7-20200525/linux-amd64/go1.14.3 
```
Both started at `5013683`. For some reason I set them both to use `--snapshot`, so it'll take a couple of hours or so for them to generate it and act 'normally' again. Doesn't hurt to do moar snapshot runs, I guess.