@holiman 2020-05-26
Here's from when the two were finished generating snapshot: https://geth-bench.ethdevops.io/d/Jpk-Be5Wk/dual-geth?orgId=1&var-exp=bench01&var-master=bench02&var-percentile=50&from=1590442980695&to=now . This PR is `bench01`, `master` is `bench02`