@holiman 2020-05-28
Oddly enough, in this portion, the `uint256` PR was ahead, but was overtaken by `master`. 
https://geth-bench.ethdevops.io/d/Jpk-Be5Wk/dual-geth?orgId=1&var-exp=bench01&var-master=bench02&var-percentile=50&from=1590538399305&to=1590591365741