@holiman 2020-05-31
Pushed a change. I may have found how/why the uint256 branch regresses: and it's due to the overhead of mem usage calls are made. It "allocates" (not via malloc, but on the stack) 1024 *32-byte at every `call`, which was not the case for master, since it previously used pointers instead. 

The `capacity` of a slice is immediately allocated. From the docs:
```
//      For example, make([]int, 0, 10) allocates an underlying array
//    of size 10 and returns a slice of length 0 and capacity 10 that is
//    backed by this underlying array.
```

Rerunning the benchmark, now with `05e5643483b7db8adfe20d1a9ca66fb5ae6a6744` (this PR) on `bench01` against `a5eee8d1dc61352c29b9800eaf96609ba4184fd6` (`master`) on `bench02`.