@holiman 2020-06-02
Here's a segment taken 24 hours into the benchmark

![Screenshot_2020-06-02 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/83494468-9fc96300-a4b6-11ea-97d0-7a66c1413554.png)

The average of `execution` is `28.4` vs `36.9`. 

Here are the last 6 hours: 

![Screenshot_2020-06-02 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/83494683-ed45d000-a4b6-11ea-9a31-63b24f98a2d1.png)