@holiman 2020-06-08
rebased